generate css with travis (#229)
* generate css with travis

Eventually this could work. We will see...

* sassc -> sass

* Travis need sudo

* use ruby-sass

* Use sass -t compress

This don't need a extra script for minify css

* delete old python script

* make deployment optional

* prepare travis switch to show this is live system

* new method including the css

Now lektor detects if the file is local available. If it is, it will use it.
If not, it will use the remote file and linking to the local font folders, because not every browser will accept the remote fonts that easy on localhost.

* Expand sass explaination

* improve lektor locale css config

lektor does not trigger a new build only if a file like main.min.css
suddenly appear or dissappear. It need a manual trigger like deleting
the cache or modify the templeates.

make will now delete the cache.